Move common OpenStack stuff to its own module.
I.e. just OpenStackBaseConnection at the moment.
