+++ libcloud/loadbalancer/drivers/rackspace.py
@@ -25,9 +25,9 @@ from libcloud.common.base import Response
 from libcloud.loadbalancer.base import LoadBalancer, Member, Driver, Algorithm
 from libcloud.loadbalancer.base import DEFAULT_ALGORITHM
 from libcloud.loadbalancer.types import State
-from libcloud.common.rackspace import (AUTH_HOST_US, AUTH_HOST_UK,
-        OpenStackBaseConnection)
-
+from libcloud.common.openstack import OpenStackBaseConnection
+from libcloud.common.rackspace import (
+        AUTH_HOST_US, AUTH_HOST_UK)
 
 class RackspaceResponse(Response):
