+++ libcloud/storage/drivers/cloudfiles.py
@@ -34,9 +34,10 @@ from libcloud.storage.types import ObjectDoesNotExistError
 from libcloud.storage.types import ObjectHashMismatchError
 from libcloud.storage.types import InvalidContainerNameError
 from libcloud.common.types import LazyList
+from libcloud.common.openstack import OpenStackBaseConnection
 
 from libcloud.common.rackspace import (
-    AUTH_HOST_US, AUTH_HOST_UK, OpenStackBaseConnection)
+    AUTH_HOST_US, AUTH_HOST_UK)
 
 CDN_HOST = 'cdn.clouddrive.com'
 API_VERSION = 'v1.0'
